Renaming StateValidator to BaseStateManager
