+++ packages/wallet/src/wallet.ts
@@ -17,11 +17,9 @@ import {
   SignedTransactionWithProof,
   Block,
   Segment,
-  SumMerkleProof,
   ChamberResult,
   ChamberResultError,
   ChamberOk,
-  SegmentChecker,
   SwapRequest,
   StateUpdate,
   OwnershipPredicate,
